+++ mobile/android/android-components/components/feature/pwa/src/main/java/mozilla/components/feature/pwa/WebAppShortcutManager.kt
@@ -59,14 +59,21 @@ class WebAppShortcutManager(
 
     /**
      * Request to create a new shortcut on the home screen.
+     * @param context The current context.
+     * @param session The session to create the shortcut for.
+     * @param overrideShortcutName (optional) The name of the shortcut
      */
-    suspend fun requestPinShortcut(context: Context, session: Session) {
+    suspend fun requestPinShortcut(
+        context: Context,
+        session: Session,
+        overrideShortcutName: String? = null
+    ) {
         if (ShortcutManagerCompat.isRequestPinShortcutSupported(context)) {
             val manifest = session.installableManifest()
             val shortcut = if (supportWebApps && manifest != null) {
                 buildWebAppShortcut(context, manifest)
             } else {
-                buildBasicShortcut(context, session)
+                buildBasicShortcut(context, session, overrideShortcutName)
             }
 
             if (shortcut != null) {
@@ -93,13 +100,17 @@ class WebAppShortcutManager(
     /**
      * Create a new basic pinned website shortcut using info from the session.
      */
-    fun buildBasicShortcut(context: Context, session: Session): ShortcutInfoCompat? {
+    fun buildBasicShortcut(
+        context: Context,
+        session: Session,
+        overrideShortcutName: String? = null
+    ): ShortcutInfoCompat {
         val shortcutIntent = Intent(Intent.ACTION_VIEW, session.url.toUri()).apply {
             `package` = context.packageName
         }
 
         val builder = ShortcutInfoCompat.Builder(context, session.url)
-            .setShortLabel(session.title.ifBlank(fallbackLabel))
+            .setShortLabel((overrideShortcutName ?: session.title).ifBlank(fallbackLabel))
             .setIntent(shortcutIntent)
 
         session.icon?.let {
@@ -112,7 +123,10 @@ class WebAppShortcutManager(
     /**
      * Create a new Progressive Web App shortcut using a web app manifest.
      */
-    suspend fun buildWebAppShortcut(context: Context, manifest: WebAppManifest): ShortcutInfoCompat? {
+    suspend fun buildWebAppShortcut(
+        context: Context,
+        manifest: WebAppManifest
+    ): ShortcutInfoCompat? {
         val shortcutIntent = Intent(context, WebAppLauncherActivity::class.java).apply {
             action = ACTION_PWA_LAUNCHER
             data = manifest.startUrl.toUri()